requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oracledb==3.4.1
pycparser==2.23
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.5
WTForms==3.2.1
ttkthemes==3.2.2
email-validator==2.2.0
matplotlib==3.8.2
Pillow==10.2.0